app/requirements.txt
azure-functions
langchain